+++ storagemarket/provider.go
@@ -2,6 +2,7 @@ package storagemarket
 
 import (
 	"context"
+	"database/sql"
 	"fmt"
 	"os"
 	"path"
@@ -27,6 +28,8 @@ import (
 )
 
 var log = logging.Logger("boost-provider")
+
+var ErrDealNotFound = fmt.Errorf("deal not found")
 
 type Config struct {
 	MaxTransferDuration time.Duration
@@ -99,6 +102,14 @@ func NewProvider(repoRoot string, dealsDB *db.DealsDB, fullnodeApi v1api.FullNod
 
 		dealHandlers: newDealHandlers(),
 	}, nil
+}
+
+func (p *Provider) Deal(ctx context.Context, dealUuid uuid.UUID) (*types.ProviderDealState, error) {
+	deal, err := p.db.ByID(ctx, dealUuid)
+	if xerrors.Is(err, sql.ErrNoRows) {
+		return nil, fmt.Errorf("getting deal %s: %w", dealUuid, ErrDealNotFound)
+	}
+	return deal, nil
 }
 
 func (p *Provider) NBytesReceived(deal *types.ProviderDealState) (int64, error) {
